consolidate ch.poole-dependencies in pom.xml files
Before 32cfae25, Graphhopper used version variables defined in the
<properties>-section of the graphhopper-parent pom.xml.
In #1906 and PR #1912 they decided to switch to maven
<dependencyManagement>.

This allows to declare versions and then use them in any child pom.xml
by only stating group- and artifact Id of the dependency.

Although the two dependencies are only used in graphhopper-core, the
versions are defined consistently in the graphhopper-parent xml.